New file

+++ hash_tables/2-key_index.c
@@ -10,5 +10,5 @@ unsigned long int hash_djb2(const unsigned char *str);
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	return (hash_djb2(key) % size);
+	return (unsigned long int)(hash_djb2(key) % size);
 }
